requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
pillow>=10.0.0
numpy>=1.24.0
librosa>=0.10.0
pdf2image>=1.16.0
reportlab>=4.0.0
python-multipart>=0.0.6
requests>=2.31.0
scipy>=1.11.0
matplotlib>=3.7.0
